+++ lib/list/refreshable_staggered_gridview.dart
@@ -1,42 +1,54 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_ahlib/list/placeholder_text.dart';
 import 'package:flutter_ahlib/list/scroll_more_controller.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 
 typedef GetNonPageDataFunction<T> = Future<List<T>> Function();
 
-class RefreshableListView<T> extends StatefulWidget {
-  const RefreshableListView({
+class RefreshableStaggeredGridView<T> extends StatefulWidget {
+  const RefreshableStaggeredGridView({
     Key key,
     @required this.data,
     @required this.getData,
     this.onStateChanged,
     this.padding,
-    this.separator,
     @required this.itemBuilder,
     this.controller,
     this.topWidget,
     this.bottomWidget,
+    this.primary,
+    this.crossAxisSpacing = 0,
+    this.mainAxisSpacing = 0,
+    @required this.crossAxisCount,
+    @required this.staggeredTileBuilder,
   })  : assert(data != null),
         assert(getData != null),
         assert(itemBuilder != null),
+        assert(crossAxisCount != null),
+        assert(staggeredTileBuilder != null),
         super(key: key);
 
   final List<T> data;
   final GetNonPageDataFunction<T> getData;
   final StateChangedCallback onStateChanged;
   final EdgeInsetsGeometry padding;
-  final Widget separator;
   final Widget Function(BuildContext, T) itemBuilder;
   final ScrollMoreController controller;
   final Widget topWidget;
   final Widget bottomWidget;
 
+  final bool primary;
+  final double crossAxisSpacing;
+  final double mainAxisSpacing;
+  final int crossAxisCount;
+  final StaggeredTile Function(int) staggeredTileBuilder;
+
   @override
-  _RefreshableListViewState<T> createState() => _RefreshableListViewState<T>();
+  _RefreshableStaggeredGridViewState<T> createState() => _RefreshableStaggeredGridViewState<T>();
 }
 
-class _RefreshableListViewState<T> extends State<RefreshableListView<T>>
-    with AutomaticKeepAliveClientMixin<RefreshableListView<T>> {
+class _RefreshableStaggeredGridViewState<T> extends State<RefreshableStaggeredGridView<T>>
+    with AutomaticKeepAliveClientMixin<RefreshableStaggeredGridView<T>> {
   @override
   bool get wantKeepAlive => true;
 
@@ -104,12 +116,16 @@ class _RefreshableListViewState<T> extends State<RefreshableListView<T>>
             widget.topWidget ?? SizedBox(height: 0),
             Expanded(
               child: Scrollbar(
-                child: ListView.separated(
+                child: StaggeredGridView.countBuilder(
                   controller: _controller,
                   physics: AlwaysScrollableScrollPhysics(),
                   padding: widget.padding,
+                  primary: widget.primary,
+                  crossAxisSpacing: widget.crossAxisSpacing,
+                  mainAxisSpacing: widget.mainAxisSpacing,
+                  crossAxisCount: widget.crossAxisCount,
+                  staggeredTileBuilder: widget.staggeredTileBuilder,
                   itemCount: widget.data.length,
-                  separatorBuilder: (c, idx) => widget.separator ?? SizedBox(height: 0),
                   itemBuilder: (c, idx) => widget.itemBuilder(c, widget.data[idx]),
                 ),
               ),